requirements.txt
wheel==0.35.1
requests==2.28.1
robotframework==5.0
zipp==3.2.0
oauthlib==3.1.0
hanging_threads==2.0.5
pywebview==3.4
pkce==1.0.3
selenium==4.8.2
chromedriver-autoinstaller==0.4.0
pyotp==2.6.0
webdrivermanager
PyYAML
robotframework-datadriver==1.7.0
robotframework-reportportal==5.0.0
reportportal-client==5.0.0